--- src/index.js
@@ -1,3 +1,0 @@
-import Plotly from "./components/Plotly.vue";
-
-export { Plotly };
